+++ tomcat/src/main/java/org/apache/coyote/http11/controller/apicontroller/RegisterApiController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller.ApiHandler;
+package org.apache.coyote.http11.controller.apicontroller;
 
 import java.io.IOException;
 import java.util.Map;
