+++ tomcat/src/main/java/org/apache/coyote/http11/controller/filecontroller/FileController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller.ApiHandler;
+package org.apache.coyote.http11.controller.filecontroller;
 
 import java.io.File;
 import java.io.IOException;
@@ -12,28 +12,31 @@ import org.apache.coyote.http11.httpmessage.request.HttpMethod;
 import org.apache.coyote.http11.httpmessage.request.HttpRequest;
 import org.apache.coyote.http11.httpmessage.response.HttpResponse;
 
-public class RegisterPageApiController extends AbstractController {
+public class FileController extends AbstractController {
 
-    private static final Pattern REGISTER_URI_PATTERN = Pattern.compile("/register");
+    private static final Pattern FILE_URI_PATTERN = Pattern.compile("/.+\\.(html|css|js|ico)");
 
     @Override
     public boolean canHandle(HttpRequest httpRequest) {
-        return httpRequest.matchRequestLine(HttpMethod.GET, REGISTER_URI_PATTERN);
+        return httpRequest.matchRequestLine(HttpMethod.GET, FILE_URI_PATTERN);
     }
 
     @Override
     public void service(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        String responseBody = getBody();
+        URL resource = getClass().getClassLoader().getResource("static" + httpRequest.getPath());
 
-        httpResponse.ok(responseBody)
-                .addHeader("Content-Type", ContentType.HTML.getValue() + ";charset=utf-8 ");
-    }
+        if (resource == null) {
+            httpResponse.notFound();
+            return;
+        }
 
-    private String getBody() throws IOException {
-        URL resource = getClass().getClassLoader().getResource("static/register.html");
         File file = new File(resource.getFile());
         Path path = file.toPath();
-        return new String(Files.readAllBytes(path));
+        String responseBody = new String(Files.readAllBytes(path));
+        ContentType contentType = ContentType.of(Files.probeContentType(path));
+
+        httpResponse.ok(responseBody)
+                .addHeader("Content-Type", contentType.getValue() + ";charset=utf-8 ");
     }
 
     @Override
