+++ tomcat/src/main/java/org/apache/coyote/http11/requestmapping/FileHandlerMapper.java
@@ -1,12 +1,12 @@
-package org.apache.coyote.http11.handlermapper;
+package org.apache.coyote.http11.requestmapping;
 
 import java.util.List;
 import org.apache.coyote.http11.controller.Controller;
-import org.apache.coyote.http11.controller.FileHandler.FileController;
-import org.apache.coyote.http11.controller.FileHandler.LoginPageHandler;
+import org.apache.coyote.http11.controller.filecontroller.FileController;
+import org.apache.coyote.http11.controller.filecontroller.LoginPageHandler;
 import org.apache.coyote.http11.httpmessage.request.HttpRequest;
 
-public class FileHandlerMapper implements HandlerMapper {
+public class FileHandlerMapper implements RequestMapper {
 
     private static final List<Controller> FILE_URI_HANDLER = List.of(new LoginPageHandler(), new FileController());
